Corrects PDO Exception claiming database not selected.
The spot configuration was referencing a database configuration key that
did not exist. The config was referencing 'database.database' where only
the following exist:

- dsn
- user
- pass

I added a fourth configuration setting, 'name', and modified both the spot
configuration and all dist configs to include that property.

A longer-term solution might be to somehow figure out how to get Eloquent (Sentry)
and Spot to share a connection or to extract configuration data from the
Eloquent connection to not have to duplicate database name in the database.dsn and
database.name properties.