Adds pre-registration check to DatabaseServiceProvider.
Before registering an instance of PDO in the service container, this
service provider will now make a preflight call to #checkConnection(..).
It queries for current databases and if the first one plucked off the list
is null, it will throw an exception with an appropriate message.

First implementation tried a query like `select 1;`, but that doesn't
produce the same error case. PDO is an odd flower for sure. The root of
this all boils down to PDO using a different DSN format than what you'd
expect working with odbc in Java-land or else. At least we're somewhat
covered here now.

References #179