Add TalkSubmission object used to represent a valid talk submission.
- Adds tests and implementation. Tests could use a refactor to reduce LoC.
- Adds two new exceptions: general validation exception and a talk submission specific one.

There is an assumption that all data is clean per requirements of request
gateway. In other words, if this object is created from an HTML form post,
data should be cleaned by HTMLPurifier for XSS before constructing the object.

Likewise, similar precautions should be taken when accepting inputs from CLI or
JSON API. The idea is that we are separating concerns of a "valid request"
from the cleaning of data.