Add second partial implementation of TalkController.
I did this to test whether or not Silex would allow us to register
controllers-as-services to get full inversion control and at the same time
not require us to move every controller over at the same time. It does, as
confirmed by next commit!

NeoTalkController will likely become the new web controller and a separate will
be created for the JSON API endpoint. They will share a dependency on Speakers
application service.

I'm also going to wrap conference configuration in an object the service is
dependent on. This will actually kill a lot of birds with one stone. Makes
testing temporal stuff a lot easier for sure! Cool.