Resolve YAML::parse deprecation notice.
See https://github.com/igorw/ConfigServiceProvider/pull/46

When the above PR is merged, we can revert this commit. The current code
is throwing errors in the test suite due to deprecation notices from Yaml::parse.
We haven't upgraded this package so no idea why we started getting notices like
this but it is what it is.

Parse method now triggers a warning if you give it a path. It expects the yaml file
contents, not a path. This patch adds a provider that extends Igor's and makes
this change. We then sub this provider into the ChainConfigProvider instead of
using SilexConfigServiceProvider's default.