Fix defaults for users table.
When using more strict MySQL settings it is not possible to add records without
specifying values for NOT NULL columns unless a default value is specified.
Phinx creates NOT NULL columns by default so many fields were giving us
problems when trying to register a user without giving a proper value
for many fields that the signup controller wasn't asking about.

This fixes the problem by altering the database schema such that the fields in
question will either allow null values or specify a default value in the case
that NULL is used or no value is provided.