Add tags to talks
Currently there is no way for users to add tags or keywords to their talks when submitting to an event. This modification adds a way so people can add such keywords or tags. This will help event organizers identifier specifics about talks and be able to group them in different ways.