Fix query for getting top rated talks
Instead of using a left join, use an inner join to join the
talks and the talk_meta tables.

This has double benefits:

 - the query is more performant as MySQL scans less rows
 - the query does not break on MySQL 5.7.5 or higher, due to
   ONLY_FULL_GROUP_BY SQL mode.

Sadly the test I add for the getTopRatedByUserId does not actually
test for this query error, because the test is currently set up to
use an in memory sqlite database.