Use models for admin dashboard controller
Changed out everything related to spot, and moved some functionality
from the talk mapper to the talk model.

Since the models get created with the new keyword we can't mock them
the normal way, so we need to add 'overload:' before the class to tell
mockery to override those as well. This also requires the use of
full namespaces instead of using them, since that messes with the magic.