Refactor viewing of talks to use Illuminate
The view action in the admin talk controller is now using illuminate
instead of spot.

This required a few changes to the Models. The TalkMeta now has a
created at const to override the standard, and a function to
override the updated_at. Might be better to add an updated_at column
in the future.

Due to the differences between spot and illuminate a few changes have
been made to the twig template.