Make the before function which adds the active function an early event
Before this, if the route the user requested wasn't defined, it would
cause an error in the 404 page, as it didn't go through the before
functions.

This commit fixes that by first of all binding it to app, instead of web
so it checks in all routes, even if it doesn't match the web, or any
route we know. It also makes it an Early event, which means that it will
fire off, even if its a 404.