Fix comments for new UI
Twig would error if some values were null if we
handed the view an object instead of an array, but we can't grab
the comments of a users when they are in array form.

So now the comments of a talk get passed to the view in a separate
variable.