Add the speaker access class
This class works the same as the admin access class, except it only
checks if a user is logged in. This should be added as middleware,
so a lot of the auth checks can be removed.

There should still be a check if the user is allowed on this specific
page. (Not editing someone else's profile etc)