Move Access checks to generic class
The Role Access class checks if a user has the required role to see
a certain page. If no role is defined it requires the admin role
( in case someone forgets to specify the needed role)

Removed the Admin and Reviewer Access classes, and their tests.
And updated the WebGateWayProvider to use the new role access class