Fix Tests
Tests needing the database have been split into 2 categories
Those that need certain data to exist, and those that remove or
heavily change the data.

The ones that just read, or change small things (ratings on a talk) use
the RefreshDatabaseTrait. This gets called before the class starts
and loads up the database fresh and new. Note that if you change entries
in a class that uses this trait you can mess up other tests.

The Tests that delete/heavily change entries use the DataBaseInteraction
trait. Which refreshes the Database back to zero before every test.

Currently thats only Model tests which have been moved to an Interaction
folder within the models, to keep them separate.