Update the talk profile
Changed the logic so now the class takes a Talk as its constructor
argument, and no longer has a with function to set the talk.

Addition of a getId function which is going to be needed in the views.

Addition of a getOtherTalks function, to get the other talks for the
view.

Changed some names to better reflect their the functions intentions

Added comments to the class

Updated tests