Implement the TalkProfile in the TalkFormatter
Instead of building a big array, the TalkFormatter now builds
talk profiles.

To make this work the Formatter + its interface needed updating.
It also required some variables in the views to be reworded, and some
updates to tests.

This removes the big mocking in the admin dashboard controller test,
and lets it use the database a bit, which turns out to be faster.

It adds a test for the reviewer dashboard controller, it updates the
tests for the TalkFormatter since that no longer returns an array.