Add a sentinel version of the account management and the user
These new versions of the account management and the user are here
to replace their old sentry versions.

Some of the functionality isn't very 'pretty' so to say, since the way
some things are handled in Sentinel are a bit different from Sentry.

The current version should be seen as an in-between, to make the change
from Sentry to Sentinel.