Separate the uploads path for production
On some production environments its not possible to have files persist
on the disk. For example the google cloud engine.

By separating the production version of the uploads path this allows
developers to still use the dev version locally, but not have to
manually edit things every time a deployment happens.

Another possibility would be to move the uploads location to the
config file, but this would make it impossible to grab files from remote
locations.