Implement one to ten rating option and move js
Split the js into a talk file, and a listeners file, and make it
more generic so both the reviewer and the admin use the same file.

Allow the option of a 1 to ten rating system and now support it in the
front-end as well.

Right now in order to change it one needs to edit the provider, but
this should be moved to a config file.