Display the airport information the speaker entered
If the speaker didn't enter the correct airport information, there would
be no display of it what so ever, and the admin wouldn't see any of the
information.

This changes that to still display the airport information the speaker
entered, even if he entered a wrong code, or just typed out the full
airport name.