Use one phpunit.xml.dist with multiple test suites
So instead of pointing to different config files we point to the test
suite that we want.

One of the big advantages of this, and the removal of the covers
annotation is that we can now get a lot better information from the
infection tests. Since those rely on the coverage data generated by
phpunit.